# Back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation",
        "contact": {
            "name": "Soporte",
            "email": "soporte@terciariourquiza.edu.ar"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/carreras": {
            "get": {
                "tags": [
                    "Carrera"
                ],
                "summary": "Obtener todas las carreras",
                "description": "Devuelve todas las carreras",
                "operationId": "getCambiosDocente",
                "responses": {
                    "200": {
                        "description": "Carreras",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Carrera"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener las carreras"
                    }
                }
            }
        },
        "/api/carreras/{id}": {
            "get": {
                "tags": [
                    "Carrera"
                ],
                "summary": "Obtener una carrera por id",
                "description": "Devuelve una carrera",
                "operationId": "getCambioDocentePorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carrera",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carrera",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Carrera"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existe la carrera"
                    },
                    "500": {
                        "description": "Error al obtener la carrera"
                    }
                }
            }
        },
        "/api/carreras/guardar": {
            "post": {
                "tags": [
                    "Carrera"
                ],
                "summary": "Guardar una carrera",
                "description": "Guardar una carrera",
                "operationId": "guardarCambioDocente",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Carrera"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carrera guardada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Carrera"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al guardar la carrera"
                    }
                }
            }
        },
        "/api/carreras/actualizar/{id}": {
            "put": {
                "tags": [
                    "Carrera"
                ],
                "summary": "Actualizar una carrera",
                "description": "Actualizar una carrera",
                "operationId": "actualizarCambioDocente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carrera",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Carrera"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carrera actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Carrera"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existe la carrera"
                    },
                    "500": {
                        "description": "Error al actualizar la carrera"
                    }
                }
            }
        },
        "/api/carreras/eliminar/{id}": {
            "delete": {
                "tags": [
                    "Carrera"
                ],
                "summary": "Eliminar una carrera",
                "description": "Eliminar una carrera",
                "operationId": "eliminarCambioDocentePorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carrera",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carrera eliminada correctamente"
                    },
                    "404": {
                        "description": "No existe la carrera"
                    },
                    "500": {
                        "description": "Error al eliminar la carrera"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Carrera": {
                "title": "Carrera",
                "description": "Esquema del objeto Carrera",
                "properties": {
                    "Id_Carrera": {
                        "description": "ID de la carrera",
                        "type": "integer"
                    },
                    "Carrera": {
                        "description": "Nombre de la carrera",
                        "type": "string"
                    },
                    "Cupo": {
                        "description": "Cupo de la carrera",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}